requirements.txt
streamlit==1.42.0
sqlalchemy==2.0.38
bcrypt==4.2.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
anthropic==0.45.0
vaderSentiment==3.3.2
nltk==3.9.1
plotly==5.24.1
streamlit-calendar==1.3.1
python-dotenv==1.0.1
streamlit-cookies-controller==0.0.4
